updated ec2 naming portion of code
I updated the naming portion of the code so that:
One- it's now a function
Two- it now checks for the value of the Key: Name to see if the name already exists. This is an improvement over the previous string method to check names where names would be considered 'invalid' if the input was unique but part of the string value of an existing name.